# public/openapi.yaml
openapi: 3.1.0
info:
  title: 점메추 API
  version: "1.0.0"
  description: |
    점심 메뉴 추천 서비스 API. [Notion 명세서](https://app.notion.com/p/API-37bf0c528644812cb80ecc60ad0ee5b4)를
    Supabase(Edge Function + Postgres + Auth + Storage) 위에 구현한 것입니다.

    ## 인증 (명세 0-2 · dual 전략)
    - **웹 브라우저**: 로그인 시 `accessToken`·`refreshToken` 이 HttpOnly 쿠키로 내려오고, 이후 자동 전송됩니다.
    - **모바일 클라이언트**: 응답 body 의 `accessToken` 을 저장한 뒤 `Authorization: Bearer` 헤더로 보냅니다.
    - **비로그인**: `uid` 쿠키(UUID)로 식별합니다. 서버가 자동 발급합니다.
    - 둘 다 있으면 **JWT 를 우선**합니다.

    `accessToken` 수명은 1시간입니다. 만료되면 `POST /auth/refresh` 로 재발급하세요 — 재로그인이 아닙니다.

    ## 에러 포맷 (명세 0-4)
    모든 에러는 `{ code, message, detail }` 단일 포맷입니다.
    `code` 는 `UPPER_SNAKE_CASE` 기계용 식별자이고, 프론트는 **이 값으로 분기**합니다.
    HTTP status 는 응답 헤더로만 표현하며 body 에 중복 기재하지 않습니다.

    ## 경로에 대하여
    명세 경로는 `/api/v1/*` 이고 실제 배포는 `/functions/v1/api/*` 입니다.
    라우터가 두 형태를 모두 받으므로 앞단에서 rewrite 만 걸면 명세 그대로 쓸 수 있습니다.
servers:
  - url: /api
    description: Vite dev 프록시 (동일 출처 — 쿠키가 그대로 동작합니다)
  - url: https://pwfevsslxkituyfktmqe.supabase.co/functions/v1/api
    description: 직접 호출 (ALLOWED_ORIGINS 시크릿 설정 필요)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: 웹 브라우저용. 로그인 시 HttpOnly 쿠키로 발급됩니다.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 모바일 클라이언트용.
  schemas:
    BudgetTier:
      type: string
      enum: [UNDER_5000, W5000_10000, W10000_15000, W15000_20000, OVER_20000]
      description: |
        예산 구간. **배열 순서가 곧 명세의 `order`** 입니다.

        | code | 범위 | label |
        |---|---|---|
        | `UNDER_5000` | 0 ~ 5,000 미만 | 5천원 미만 |
        | `W5000_10000` | 5,000 ~ 10,000 미만 | 5천원대 |
        | `W10000_15000` | 10,000 ~ 15,000 미만 | 1만원대 |
        | `W15000_20000` | 15,000 ~ 20,000 미만 | 1만5천원대 |
        | `OVER_20000` | 20,000 이상 | 2만원 이상 |

        **역할에 따라 해석이 다릅니다.**
        메뉴 속성일 때는 그 메뉴가 속한 버킷 하나, 추천 필터일 때는 지불 가능 **상한**입니다.
      example: W5000_10000
    Category:
      type: object
      required: [id, name]
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: 한식 }
    Menu:
      type: object
      description: 추천 응답에 실리는 메뉴 객체.
      required: [id, name, budgetTier, categories, minPeople, maxPeople]
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: 김치찌개 }
        imageUrl:
          type: [string, "null"]
          example: null
        budgetTier: { $ref: "#/components/schemas/BudgetTier" }
        categories:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        minPeople: { type: integer, example: 1 }
        maxPeople: { type: integer, example: 4 }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 프론트는 이 값으로 분기합니다.
          example: DUPLICATE_MENU
        message: { type: string, example: 이미 등록된 메뉴입니다. }
        detail:
          type: object
          description: 위반 필드 목록, 후보 배열 등 구조화된 부가정보. 없으면 `{}`.
          example: {}
  responses:
...
    ValidationError:
      description: 유효성 위반
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            code: VALIDATION_ERROR
            message: 입력값을 다시 확인해 주세요.
            detail: { fields: [budgetTier] }
paths:
...
  /recommend:
    get:
      tags: [추천]
      summary: 메뉴 추천
      description: |
        조건에 맞는 메뉴를 **한 개** 무작위로 반환하고 추천 이력을 기록합니다.

        비로그인 상태이고 `uid` 쿠키가 없으면 서버가 새로 발급해 내려줍니다.
      security:
        - {}
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: people
          in: query
          required: false
          schema: { type: integer, minimum: 1, example: 2 }
          description: 함께 먹는 실제 인원수. `minPeople ≤ people ≤ maxPeople` 인 메뉴를 매칭합니다.
        - name: budgetTier
          in: query
          required: false
          schema: { $ref: "#/components/schemas/BudgetTier" }
          description: "**상한으로 해석**합니다. `W10000_15000` 을 주면 그 이하 티어가 모두 후보입니다."
        - name: categoryIds
          in: query
          required: false
          explode: true
          schema:
            type: array
            items: { type: integer, format: int64 }
          description: |
            반복 파라미터로 보냅니다 (`categoryIds=1&categoryIds=3`).
            여러 개면 **OR 매칭** — 하나라도 해당하면 후보입니다.
      responses:
        "200":
          description: 추천 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendationId: { type: integer, format: int64, example: 1001 }
                  menu: { $ref: "#/components/schemas/Menu" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404":
          description: |
            조건에 맞는 메뉴 없음. **예외가 아니라 정상 흐름입니다.**
            단일 메뉴를 반환하는 API 라 "빈 배열 200" 으로 표현할 자리가 없어 404 를 씁니다.
            프론트는 에러 토스트가 아니라 **필터 완화를 권하는 안내 UI** 를 띄우세요.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                code: NO_MENU_FOUND
                message: 조건에 맞는 메뉴가 없어요. 필터를 넓혀 보세요.
                detail: {}
  /recommendations/{recommendationId}/action:
    post:
      tags: [추천]
      summary: 추천 행동 기록 저장
      description: 화면 ② "결정" / "다시" 버튼 클릭 시 호출합니다. Response Body 없음.
      security:
        - {}
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: recommendationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: 추천 API 응답의 `recommendationId`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [chosen, skipped]
                  description: "`chosen` = 결정 버튼, `skipped` = 다시 버튼"
            example: { action: chosen }
      responses:
        "200": { description: 기록됨 (본문 없음) }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404":
          description: |
            존재하지 않는 `recommendationId` 와 **남의 기록**을 동일하게 404 로 처리합니다.
            403 으로 구분하면 존재 여부가 노출되기 때문입니다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { code: RECOMMENDATION_NOT_FOUND, message: 추천 기록을 찾을 수 없습니다., detail: {} }
...